+++ demos/Platform/Ninja.py
@@ -10,4 +10,4 @@ class Ninja(Sprite):
         self.setOnEvents(LittleGameEngine.E_ON_UPDATE)
 
     def onUpdate(self, dt):
-        self.nextShape(dt, 0.050)
+        self.nextImage(dt, 0.050)