+++ demos/Simple 02/Demo01_TheWorld.py
@@ -36,8 +36,9 @@ class TheWorld():
 
         # agregamos el icono del sonido
         mute = Sprite("mute", (8, 3), "mute")
-        mute.setShape("mute", 1)
+        mute.setImage("mute", 1)
         self.lge.addGObjectGUI(mute)
+        self.isMute = False
 
         # agregamos al heroe
         heroe = Sprite("heroe", (226, 254), "Heroe")
@@ -74,12 +75,12 @@ class TheWorld():
             mute = self.lge.getGObject("mute")
             r = mute.getRectangle()
             if(r.contains(mx, my)):
-                idx = mute.getCurrentIdx()
-                if(idx == 1):
-                    self.lge.setSoundVolume("fondo", 0)
-                else:
+                if(self.isMute):
                     self.lge.setSoundVolume("fondo", 50)
-                mute.nextShape()
+                else:
+                    self.lge.setSoundVolume("fondo", 0)
+                self.isMute = not self.isMute
+                mute.nextImage()
 
     # main loop
     def run(self):